backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==13.0
google-genai>=1.20.0
google-cloud-aiplatform==1.70.0
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.0
google-cloud-bigquery==3.25.0
google-cloud-secret-manager==2.20.0
google-cloud-texttospeech==2.17.0
google-cloud-discoveryengine==0.13.0
pymupdf==1.24.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.9.0
httpx==0.28.1
pytest==8.3.0
pytest-asyncio==0.24.0
reportlab==4.0.0
numpy==1.26.0
qrcode[pil]==7.4.2
bcrypt==4.2.0
redis==5.0.8
slowapi==0.1.9
